# LagacyItem fix on 20250418 - fix AntiqueKatana
