+++ core/src/main/assets/messages/items/items_zh.properties
@@ -1589,6 +1589,17 @@ items.lagacyitem.dualdueldaggers.desc=若你使用这把武器造成的连续伤
 items.weapon.melee.meleeweapon.stats_knownx=这件_传承的近战武器_可以造成_%1$d~%2$d点伤害_，并且需要_%3$d点力量_来正常使用。
 items.weapon.melee.meleeweapon.stats_unknownx=一般情况下这件传承的近战武器可以造成_%1$d~%2$d点伤害_，并且需要_%3$d点力量_来正常使用。
 
+items.lagacyitem.antiquekatana.name=古武士刀
+items.lagacyitem.antiquekatana.desc=这把古老的武士刀在无数次的战斗后已经被损坏，尽管它几乎是把废铜烂铁，可它仍隐隐约约地闪现出奇怪的力量。
+
+items.lagacyitem.muramasa.name=村正
+items.lagacyitem.muramasa.desc=当使用此武器攻击命中后，给予被攻击者5回合狂乱；当自身攻击范围内存在可攻击敌人时，将无法进行除攻击外的任何操作\n\n 这把武器对野兽有35%的伤害加成 \n\n 由一位在地牢深处迷失自我的著名铁匠锻造而成的最终之作，他扭曲的灵魂潜藏在刀刃之中。传闻称使用这把利刃的人会因为寄居在刀刃里宿主的嗜血意志而变得狂暴从而攻击盟友与同类。
+
+items.lagacyitem.masamune.name=正宗
+items.lagacyitem.masamune.desc=当装备此传承武器时，角色视为始终拥有祝福状态，并免疫晕眩和幻惑\n\n 这把武器对恶魔和亡灵有35%的伤害加成 \n\n 一把由著名的铁匠经过多年锻造出来的最终之刃，它被创造的初衷是为了斩尽一切邪秽之物从而保护无辜。
+
+
+
 ###weapon curses
 items.weapon.curses.annoying.name=喧闹%s
 items.weapon.curses.annoying.msg_1=嘿，头儿！我们要去打谁！
