+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/custom/testmode/generator/LazyTest.java
@@ -10,6 +10,7 @@ import com.shatteredpixel.shatteredpixeldungeon.items.armor.PlateArmor;
 import com.shatteredpixel.shatteredpixeldungeon.items.bombs.Bomb;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.Food;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.Pasty;
+import com.shatteredpixel.shatteredpixeldungeon.items.lagacyItem.AntiqueKatana;
 import com.shatteredpixel.shatteredpixeldungeon.items.potions.AlchemicalCatalyst;
 import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfExperience;
 import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfFrost;
@@ -137,6 +138,8 @@ public class LazyTest extends TestGenerator {
             new StoneOfBlink().quantity(100).collect();
 
             new StoneOfAugmentation().quantity(100).collect();
+
+            new AntiqueKatana().identify().collect();
 
             new Blindweed.Seed().quantity(100).identify().collect();
             new Mageroyal.Seed().quantity(100).identify().collect();
